Refactor: Extract Card action functions to hook
Again, not really sure if this is better.
